--- plugins/groovy/groovy-psi/resources/intentionDescriptions/InferMethodArgumentsIntention/description.html
@@ -1,6 +1,0 @@
-<!-- Copyright 2000-2019 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file. -->
-<html>
-<body>
-Adds explicit types of method arguments.
-</body>
-</html>
