+++ plugins/groovy/groovy-psi/resources/intentionDescriptions/InferMethodParametersTypesIntention/after.groovy.template
