+++ plugins/groovy/groovy-psi/resources/intentionDescriptions/InferMethodParametersTypesIntention/before.groovy.template
