+++ plugins/groovy/groovy-psi/resources/intentionDescriptions/InferMethodParametersTypesIntention/description.html
@@ -1,0 +1,5 @@
+<html>
+<body>
+Adds explicit types for method parameters types.
+</body>
+</html>
