+++ plugins/groovy/groovy-psi/src/org/jetbrains/plugins/groovy/intentions/GroovyIntentionsBundle.properties
@@ -144,8 +144,8 @@ convert.from.geese.braces.intention.name=Format braces in separate lines
 convert.from.geese.braces.intention.family.name=Format braces in separate lines
 add.return.type=Add return type
 add.return.type.to.method.declaration=Add return type to method declaration
-infer.method.arguments=Infer method arguments
-infer.method.arguments.for.method.declaration=Infer method arguments for method declaration
+infer.method.parameters.types=Infer method parameters types
+infer.method.parameters.types.for.method.declaration=Infer method parameters types for method declaration
 replace.with.wrapper=Replace with {0}
 replace.primitive.type.with.wrapper=Replace primitive type with wrapper
 split.into.declaration.and.assignment=Split into declaration and assignment
