+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/CustomInheritance.groovy
