+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/CustomInheritance_after.groovy
