+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/DeepEqualityOfWildcards.groovy
