+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/DeepEqualityOfWildcards_after.groovy
