+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/DependencyOnTypeParameter.groovy
