+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/DependencyOnTypeParameter_after.groovy
