+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/InferWildcardsDependencies.groovy
