+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/InferWildcardsDependencies_after.groovy
