+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/NoInferenceInMethodBody.groovy
