+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/NoInferenceWithoutNontypedArguments.groovy
