+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/NoResolveForTypeParameters.groovy
