+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/NoResolveForTypeParameters_after.groovy
