+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/OperatorInference.groovy
