+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/OperatorInference_after.groovy
