+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/RelatedWildcards.groovy
