+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/RelatedWildcards_after.groovy
