+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/ResolvedMethodCallInside.groovy
