+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/ResolvedMethodCallInside_after.groovy
