+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/SearchEverywhere.groovy
