+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/SearchEverywhere_after.groovy
