+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/SeveralArguments.groovy
