+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/SeveralArguments_after.groovy
