+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/StringInference.groovy
