+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/StringInference_after.groovy
