+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/UpperBoundByTypeParameter.groovy
