+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/UpperBoundByTypeParameter_after.groovy
